refactor: rename client impl query package to request
