+++ clients/java/src/main/java/io/camunda/client/impl/CamundaClientImpl.java
@@ -169,17 +169,17 @@ import io.camunda.client.impl.fetch.UserTaskGetRequestImpl;
 import io.camunda.client.impl.fetch.VariableGetRequestImpl;
 import io.camunda.client.impl.http.HttpClient;
 import io.camunda.client.impl.http.HttpClientFactory;
-import io.camunda.client.impl.search.query.AdHocSubprocessActivitySearchRequestImpl;
-import io.camunda.client.impl.search.query.DecisionDefinitionSearchRequestImpl;
-import io.camunda.client.impl.search.query.DecisionInstanceSearchRequestImpl;
-import io.camunda.client.impl.search.query.DecisionRequirementsSearchRequestImpl;
-import io.camunda.client.impl.search.query.FlowNodeInstanceSearchRequestImpl;
-import io.camunda.client.impl.search.query.IncidentSearchRequestImpl;
-import io.camunda.client.impl.search.query.ProcessDefinitionSearchRequestImpl;
-import io.camunda.client.impl.search.query.ProcessInstanceSearchRequestImpl;
-import io.camunda.client.impl.search.query.UserTaskSearchRequestImpl;
-import io.camunda.client.impl.search.query.UserTaskVariableSearchRequestImpl;
-import io.camunda.client.impl.search.query.VariableSearchRequestImpl;
+import io.camunda.client.impl.search.request.AdHocSubprocessActivitySearchRequestImpl;
+import io.camunda.client.impl.search.request.DecisionDefinitionSearchRequestImpl;
+import io.camunda.client.impl.search.request.DecisionInstanceSearchRequestImpl;
+import io.camunda.client.impl.search.request.DecisionRequirementsSearchRequestImpl;
+import io.camunda.client.impl.search.request.FlowNodeInstanceSearchRequestImpl;
+import io.camunda.client.impl.search.request.IncidentSearchRequestImpl;
+import io.camunda.client.impl.search.request.ProcessDefinitionSearchRequestImpl;
+import io.camunda.client.impl.search.request.ProcessInstanceSearchRequestImpl;
+import io.camunda.client.impl.search.request.UserTaskSearchRequestImpl;
+import io.camunda.client.impl.search.request.UserTaskVariableSearchRequestImpl;
+import io.camunda.client.impl.search.request.VariableSearchRequestImpl;
 import io.camunda.client.impl.util.ExecutorResource;
 import io.camunda.client.impl.util.VersionUtil;
 import io.camunda.client.impl.worker.JobClientImpl;
