+++ clients/java/src/main/java/io/camunda/client/impl/fetch/ProcessDefinitionGetRequestImpl.java
@@ -21,7 +21,7 @@ import io.camunda.client.api.fetch.ProcessDefinitionGetRequest;
 import io.camunda.client.api.search.response.ProcessDefinition;
 import io.camunda.client.impl.http.HttpCamundaFuture;
 import io.camunda.client.impl.http.HttpClient;
-import io.camunda.client.impl.search.SearchResponseMapper;
+import io.camunda.client.impl.search.response.SearchResponseMapper;
 import io.camunda.client.protocol.rest.ProcessDefinitionResult;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
