+++ clients/java/src/main/java/io/camunda/client/impl/search/filter/DecisionDefinitionFilterImpl.java
@@ -16,7 +16,7 @@
 package io.camunda.client.impl.search.filter;
 
 import io.camunda.client.api.search.filter.DecisionDefinitionFilter;
-import io.camunda.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.client.impl.search.request.TypedSearchRequestPropertyProvider;
 import io.camunda.client.impl.util.ParseUtil;
 
 public class DecisionDefinitionFilterImpl
