+++ clients/java/src/main/java/io/camunda/client/impl/search/filter/FlownodeInstanceFilterImpl.java
@@ -18,7 +18,7 @@ package io.camunda.client.impl.search.filter;
 import io.camunda.client.api.search.filter.FlownodeInstanceFilter;
 import io.camunda.client.api.search.response.FlowNodeInstanceState;
 import io.camunda.client.api.search.response.FlowNodeInstanceType;
-import io.camunda.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.client.impl.search.request.TypedSearchRequestPropertyProvider;
 import io.camunda.client.impl.util.ParseUtil;
 
 public class FlownodeInstanceFilterImpl
