+++ clients/java/src/main/java/io/camunda/client/impl/search/filter/UserTaskVariableFilterImpl.java
@@ -17,8 +17,8 @@ package io.camunda.client.impl.search.filter;
 
 import io.camunda.client.api.search.filter.UserTaskVariableFilter;
 import io.camunda.client.api.search.filter.builder.StringProperty;
-import io.camunda.client.impl.search.TypedSearchRequestPropertyProvider;
 import io.camunda.client.impl.search.filter.builder.StringPropertyImpl;
+import io.camunda.client.impl.search.request.TypedSearchRequestPropertyProvider;
 import io.camunda.client.protocol.rest.VariableUserTaskFilterRequest;
 import java.util.function.Consumer;
