+++ clients/java/src/main/java/io/camunda/client/impl/search/request/AdHocSubprocessActivitySearchRequestImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.impl.search.query;
+package io.camunda.client.impl.search.request;
 
 import static io.camunda.client.api.search.request.SearchRequestBuilders.adHocSubprocessActivityFilter;
