+++ clients/java/src/main/java/io/camunda/client/impl/search/request/DecisionDefinitionSearchRequestImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.impl.search.query;
+package io.camunda.client.impl.search.request;
 
 import static io.camunda.client.api.search.request.SearchRequestBuilders.decisionDefinitionFilter;
 import static io.camunda.client.api.search.request.SearchRequestBuilders.decisionDefinitionSort;
@@ -30,9 +30,7 @@ import io.camunda.client.api.search.response.SearchResponse;
 import io.camunda.client.api.search.sort.DecisionDefinitionSort;
 import io.camunda.client.impl.http.HttpCamundaFuture;
 import io.camunda.client.impl.http.HttpClient;
-import io.camunda.client.impl.search.SearchRequestSortMapper;
-import io.camunda.client.impl.search.SearchResponseMapper;
-import io.camunda.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.client.impl.search.response.SearchResponseMapper;
 import io.camunda.client.protocol.rest.DecisionDefinitionSearchQueryResult;
 import io.camunda.client.protocol.rest.SearchQueryPageRequest;
 import java.time.Duration;
@@ -74,7 +72,7 @@ public class DecisionDefinitionSearchRequestImpl
 
   @Override
   public DecisionDefinitionSearchRequest sort(final DecisionDefinitionSort value) {
-    final List<io.camunda.client.impl.search.SearchRequestSort> sorting =
+    final List<io.camunda.client.impl.search.request.SearchRequestSort> sorting =
         provideSearchRequestProperty(value);
     request.setSort(SearchRequestSortMapper.toDecisionDefinitionSearchQuerySortRequest(sorting));
     return this;
