+++ clients/java/src/main/java/io/camunda/client/impl/search/request/UserTaskVariableSearchRequestImpl.java
@@ -13,49 +13,50 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.impl.search.query;
+package io.camunda.client.impl.search.request;
 
 import static io.camunda.client.api.search.request.SearchRequestBuilders.searchRequestPage;
-import static io.camunda.client.api.search.request.SearchRequestBuilders.variableFilter;
+import static io.camunda.client.api.search.request.SearchRequestBuilders.userTaskVariableFilter;
 import static io.camunda.client.api.search.request.SearchRequestBuilders.variableSort;
 
 import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
-import io.camunda.client.api.search.filter.VariableFilter;
+import io.camunda.client.api.search.filter.UserTaskVariableFilter;
 import io.camunda.client.api.search.request.FinalSearchRequestStep;
 import io.camunda.client.api.search.request.SearchRequestPage;
-import io.camunda.client.api.search.request.VariableSearchRequest;
+import io.camunda.client.api.search.request.UserTaskVariableSearchRequest;
 import io.camunda.client.api.search.response.SearchResponse;
 import io.camunda.client.api.search.response.Variable;
 import io.camunda.client.api.search.sort.VariableSort;
 import io.camunda.client.impl.http.HttpCamundaFuture;
 import io.camunda.client.impl.http.HttpClient;
-import io.camunda.client.impl.search.SearchRequestPageImpl;
-import io.camunda.client.impl.search.SearchRequestSortMapper;
-import io.camunda.client.impl.search.SearchResponseMapper;
-import io.camunda.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.client.impl.search.response.SearchResponseMapper;
 import io.camunda.client.impl.search.sort.VariableSortImpl;
-import io.camunda.client.protocol.rest.VariableSearchQuery;
+import io.camunda.client.protocol.rest.UserTaskVariableSearchQueryRequest;
 import io.camunda.client.protocol.rest.VariableSearchQueryResult;
+import io.camunda.client.protocol.rest.VariableUserTaskFilterRequest;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import org.apache.hc.client5.http.config.RequestConfig;
 
-public class VariableSearchRequestImpl
-    extends TypedSearchRequestPropertyProvider<VariableSearchQuery>
-    implements VariableSearchRequest {
+public class UserTaskVariableSearchRequestImpl
+    extends TypedSearchRequestPropertyProvider<UserTaskVariableSearchQueryRequest>
+    implements UserTaskVariableSearchRequest {
 
-  private final VariableSearchQuery request;
-  private final JsonMapper jsonMapper;
+  private final UserTaskVariableSearchQueryRequest request;
   private final HttpClient httpClient;
   private final RequestConfig.Builder httpRequestConfig;
+  private final JsonMapper jsonMapper;
+  private final long userTaskKey;
 
-  public VariableSearchRequestImpl(final HttpClient httpClient, final JsonMapper jsonMapper) {
-    request = new VariableSearchQuery();
-    this.jsonMapper = jsonMapper;
+  public UserTaskVariableSearchRequestImpl(
+      final HttpClient httpClient, final JsonMapper jsonMapper, final long userTaskKey) {
     this.httpClient = httpClient;
+    this.jsonMapper = jsonMapper;
+    this.userTaskKey = userTaskKey;
     httpRequestConfig = httpClient.newRequestConfig();
+    request = new UserTaskVariableSearchQueryRequest();
   }
 
   @Override
@@ -68,7 +69,7 @@ public class VariableSearchRequestImpl
   public CamundaFuture<SearchResponse<Variable>> send() {
     final HttpCamundaFuture<SearchResponse<Variable>> result = new HttpCamundaFuture<>();
     httpClient.post(
-        "/variables/search",
+        String.format("/user-tasks/%d/variables/search", userTaskKey),
         jsonMapper.toJson(request),
         httpRequestConfig.build(),
         VariableSearchQueryResult.class,
@@ -78,46 +79,45 @@ public class VariableSearchRequestImpl
   }
 
   @Override
-  public VariableSearchRequest filter(final VariableFilter value) {
-    final io.camunda.client.protocol.rest.VariableFilter filter =
-        provideSearchRequestProperty(value);
+  public UserTaskVariableSearchRequest filter(final UserTaskVariableFilter value) {
+    final VariableUserTaskFilterRequest filter = provideSearchRequestProperty(value);
     request.setFilter(filter);
     return this;
   }
 
   @Override
-  public VariableSearchRequest filter(final Consumer<VariableFilter> fn) {
-    return filter(variableFilter(fn));
+  public UserTaskVariableSearchRequest filter(final Consumer<UserTaskVariableFilter> fn) {
+    return filter(userTaskVariableFilter(fn));
   }
 
   @Override
-  public VariableSearchRequest sort(final VariableSort value) {
+  public UserTaskVariableSearchRequest sort(final VariableSort value) {
     final VariableSortImpl sorting = (VariableSortImpl) value;
     request.setSort(
-        SearchRequestSortMapper.toVariableSearchQuerySortRequest(
+        SearchRequestSortMapper.toUserTaskVariableSearchQuerySortRequest(
             sorting.getSearchRequestProperty()));
     return this;
   }
 
   @Override
-  public VariableSearchRequest sort(final Consumer<VariableSort> fn) {
+  public UserTaskVariableSearchRequest sort(final Consumer<VariableSort> fn) {
     return sort(variableSort(fn));
   }
 
   @Override
-  public VariableSearchRequest page(final SearchRequestPage value) {
+  public UserTaskVariableSearchRequest page(final SearchRequestPage value) {
     final SearchRequestPageImpl page = (SearchRequestPageImpl) value;
     request.setPage(page.getSearchRequestProperty());
     return this;
   }
 
   @Override
-  public VariableSearchRequest page(final Consumer<SearchRequestPage> fn) {
+  public UserTaskVariableSearchRequest page(final Consumer<SearchRequestPage> fn) {
     return page(searchRequestPage(fn));
   }
 
   @Override
-  protected VariableSearchQuery getSearchRequestProperty() {
+  protected UserTaskVariableSearchQueryRequest getSearchRequestProperty() {
     return request;
   }
 }
