+++ clients/java/src/main/java/io/camunda/client/impl/search/response/SearchResponseMapper.java
@@ -1,19 +1,11 @@
 /*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.client.impl.search;
+package io.camunda.client.impl.search.response;
 
 import io.camunda.client.api.JsonMapper;
 import io.camunda.client.api.search.response.DecisionDefinition;
@@ -27,17 +19,6 @@ import io.camunda.client.api.search.response.SearchResponse;
 import io.camunda.client.api.search.response.SearchResponsePage;
 import io.camunda.client.api.search.response.UserTask;
 import io.camunda.client.api.search.response.Variable;
-import io.camunda.client.impl.search.response.DecisionDefinitionImpl;
-import io.camunda.client.impl.search.response.DecisionInstanceImpl;
-import io.camunda.client.impl.search.response.DecisionRequirementsImpl;
-import io.camunda.client.impl.search.response.FlowNodeInstanceImpl;
-import io.camunda.client.impl.search.response.IncidentImpl;
-import io.camunda.client.impl.search.response.ProcessDefinitionImpl;
-import io.camunda.client.impl.search.response.ProcessInstanceImpl;
-import io.camunda.client.impl.search.response.SearchResponseImpl;
-import io.camunda.client.impl.search.response.SearchResponsePageImpl;
-import io.camunda.client.impl.search.response.UserTaskImpl;
-import io.camunda.client.impl.search.response.VariableImpl;
 import io.camunda.client.protocol.rest.*;
 import java.util.Collections;
 import java.util.List;
