+++ clients/java/src/main/java/io/camunda/client/impl/search/sort/FlownodeInstanceSortImpl.java
@@ -16,7 +16,7 @@
 package io.camunda.client.impl.search.sort;
 
 import io.camunda.client.api.search.sort.FlownodeInstanceSort;
-import io.camunda.client.impl.search.query.SearchRequestSortBase;
+import io.camunda.client.impl.search.request.SearchRequestSortBase;
 
 public class FlownodeInstanceSortImpl extends SearchRequestSortBase<FlownodeInstanceSort>
     implements FlownodeInstanceSort {
