+++ clients/java/src/main/java/io/camunda/client/impl/search/sort/VariableSortImpl.java
@@ -16,7 +16,7 @@
 package io.camunda.client.impl.search.sort;
 
 import io.camunda.client.api.search.sort.VariableSort;
-import io.camunda.client.impl.search.query.SearchRequestSortBase;
+import io.camunda.client.impl.search.request.SearchRequestSortBase;
 
 public class VariableSortImpl extends SearchRequestSortBase<VariableSort> implements VariableSort {
