+++ clients/java/src/test/java/io/camunda/client/flownodeinstance/FlownodeInstanceTest.java
@@ -23,8 +23,8 @@ import com.github.tomakehurst.wiremock.http.RequestMethod;
 import com.github.tomakehurst.wiremock.verification.LoggedRequest;
 import io.camunda.client.api.search.response.FlowNodeInstanceState;
 import io.camunda.client.api.search.response.FlowNodeInstanceType;
-import io.camunda.client.impl.search.SearchRequestSort;
-import io.camunda.client.impl.search.SearchRequestSortMapper;
+import io.camunda.client.impl.search.request.SearchRequestSort;
+import io.camunda.client.impl.search.request.SearchRequestSortMapper;
 import io.camunda.client.protocol.rest.*;
 import io.camunda.client.util.ClientRestTest;
 import java.util.List;
