+++ clients/java/src/test/java/io/camunda/client/process/QueryProcessInstanceTest.java
@@ -21,9 +21,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.github.tomakehurst.wiremock.http.RequestMethod;
 import com.github.tomakehurst.wiremock.verification.LoggedRequest;
 import io.camunda.client.api.search.response.ProcessInstanceState;
-import io.camunda.client.impl.search.SearchRequestSort;
-import io.camunda.client.impl.search.SearchRequestSortMapper;
 import io.camunda.client.impl.search.filter.builder.StringPropertyImpl;
+import io.camunda.client.impl.search.request.SearchRequestSort;
+import io.camunda.client.impl.search.request.SearchRequestSortMapper;
 import io.camunda.client.protocol.rest.*;
 import io.camunda.client.util.ClientRestTest;
 import io.camunda.client.util.RestGatewayService;
